Brower Verify/requirements.txt
playwright==1.41.0
httpx==0.24.1
PyYAML==6.0
user-agents==2.2.0
tldextract==3.4.0
fastapi==0.115.0
uvicorn==0.30.6
Jinja2==3.1.4

